% Generated by roxygen2: do not edit by hand
\name{safer_recall}
\alias{safer_recall}
\title{Get a single recall by recall number}
\usage{
safer_recall(recall_number)
}
\arguments{
\item{recall_number}{Recall number (e.g. "24090").}
}
\value{
tibble with one row
}
\description{
Get a single recall by recall number
}
